requirements.txt
discord.py
python-dotenv
piper-tts
pynacl
